Move determine_kwargs and KeywordParameters to SDK DecoratedOperator
Move `KeywordParameters` class and `determine_kwargs()` from
`airflow.utils.operator_helpers` to `task-sdk/src/airflow/sdk/bases/decorator.py`,
next to `DecoratedOperator` which owns `python_callable` and `op_args`.

Add a `determine_kwargs()` instance method on `DecoratedOperator` (mirrors
the existing pattern on `PythonOperator`).

All 6 decorated operators now import the standalone `determine_kwargs` from
`common.compat.sdk` instead of `airflow.utils.operator_helpers`. Using the
compat layer (rather than the instance method) handles the case where a newer
provider runs against an older Task SDK whose `DecoratedOperator` lacks the
method. The compat fallback resolves to `airflow.utils.operator_helpers` on
older Airflow versions.

Other changes:
- `operator_helpers.py` re-exports via `add_deprecated_classes` so existing
  callers get a deprecation warning pointing to the SDK location
- `PythonOperator` imports `KeywordParameters` from compat SDK
- Affected provider pyproject.toml files marked with `# use next version`
  for the common-compat dependency
- Cleaned up stale `serializing()` reference in `KeywordParameters` docstring
- Removed redundant local `import itertools` (already at module level)
